dbus: add an external param `waitOnDelay` to handle service unavailable
Sometimes the D-Bus service is unavailable at bootstraping the os, so
the client will get a connection with NULL, and the following operations
with this connection would make failures even abortions.

This patch intros a synchronous wait, that the user could gives a max
timeout, within that, ShadowNode attempts to connect and check the following
errors:

- org.freedesktop.DBus.Error.FileNotFound
- org.freedesktop.DBus.Error.Failed

If one of them, client would sleep 100ms and retry until the current time
is over the given max timeout. Otherwise it breaks the loop for other
errors like "Address does not contain a colon".